:recycle:refactor: 분리한 컴포넌트를 return 할 수 있도록 변경하여 React처럼 변경 및 utils 함수 분리 처리
